# backend/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/auth/sendCode":{"post":{"operationId":"AuthController_sendCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CodeRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatus"}}}}},"tags":["Authentication"]}},"/auth/verifyCode":{"post":{"operationId":"AuthController_verifyCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CodeVerificationRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}},"tags":["Authentication"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/orders/all":{"get":{"operationId":"OrdersController_all","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}}},"tags":["Orders"]}},"/orders/{id}":{"get":{"operationId":"OrdersController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}}},"tags":["Orders"]}},"/all":{"get":{"operationId":"StoresController_all","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Store"}}}}}},"tags":["Stores"]}},"/availablePickups":{"post":{"operationId":"StoresController_availablePickups","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailablePickupsRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PickupLocation"}}}}}},"tags":["Stores"]}},"/book":{"post":{"operationId":"StoresController_book","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookPickupsRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OperationStatus"}}}}}},"tags":["Stores"]}},"/schedule":{"get":{"operationId":"ScheduleController_all","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Schedule"}}}}}},"tags":["schedule"]}},"/schedule/{id}":{"get":{"operationId":"ScheduleController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Schedule"}}}}},"tags":["schedule"]},"delete":{"operationId":"ScheduleController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["schedule"]}}},"info":{"title":"eCommerce Pickup Service","description":"The eCommerce Pickup Service API description","version":"1.0","contact":{}},"tags":[{"name":"","description":""}],"servers":[{"url":"http://localhost:50000"}],"components":{"schemas":{"CodeRequest":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"OperationStatus":{"type":"object","properties":{"status":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}}},"required":["status","errors"]},"CodeVerificationRequest":{"type":"object","properties":{"code":{"type":"string"},"email":{"type":"string"}},"required":["code","email"]},"User":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"is_admin":{"type":"boolean"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","email","name","address","phone","is_admin","created_at","updated_at"]},"AuthResponse":{"type":"object","properties":{"status":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"user":{"$ref":"#/components/schemas/User"}},"required":["status","errors","user"]},"PickupLocation":{"type":"object","properties":{"id":{"type":"number"},"sap_id":{"type":"number"},"parcel_size":{"type":"string"},"name":{"type":"string"}},"required":["id","sap_id","parcel_size","name"]},"Schedule":{"type":"object","properties":{"pickup_location":{"$ref":"#/components/schemas/PickupLocation"},"start_time":{"format":"date-time","type":"string"},"end_time":{"format":"date-time","type":"string"}},"required":["pickup_location","start_time","end_time"]},"Order":{"type":"object","properties":{"sap_id":{"type":"number"},"preparation_time":{"type":"string"},"parcel_size":{"type":"string"},"entries":{"type":"string"},"schedule":{"$ref":"#/components/schemas/Schedule"}},"required":["sap_id","preparation_time","parcel_size","entries","schedule"]},"Store":{"type":"object","properties":{"name":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"openingHours":{"type":"string"},"pickup_locations":{"type":"array","items":{"$ref":"#/components/schemas/PickupLocation"}},"employees":{"type":"array","items":{"type":"array"}}},"required":["name","phone","address","openingHours","pickup_locations","employees"]},"AvailablePickupsRequest":{"type":"object","properties":{"order_id":{"type":"number"},"store_id":{"type":"number"},"desired_time":{"format":"date-time","type":"string"}},"required":["order_id","store_id","desired_time"]},"BookPickupsRequest":{"type":"object","properties":{"order_id":{"type":"number"},"store_id":{"type":"number"},"pickup_location_id":{"type":"number"},"desired_time":{"format":"date-time","type":"string"}},"required":["order_id","store_id","pickup_location_id","desired_time"]}}}}